+++ frontend/src/domains/atoms/Web3FlagAtom.tsx
@@ -1,0 +1,3 @@
+import { atom } from "jotai";
+
+export const Web3FlagAtom = atom<boolean>(true);
